requirements/base.txt
# Base requirements for Claude PM Framework
# Core dependencies for production use

# Async HTTP and networking
aiohttp>=3.9.0
httpx>=0.25.0

# CLI and command-line tools
click>=8.1.0
typer>=0.9.0
rich>=13.7.0

# Data validation and parsing
pydantic>=2.5.0
pyyaml>=6.0.1
toml>=0.10.2
tomli>=2.0.0
python-dotenv>=1.0.0

# System monitoring
psutil>=5.9.0

# AI and memory integration
openai>=1.0.0
requests>=2.31.0

# Utilities
pathspec>=0.11.0

# Ticketing and issue tracking
ai-trackdown-pytools==1.1.0

# Markdown parsing
python-frontmatter>=1.0.0
mistune>=3.0.0

# Code analysis and parsing
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-typescript>=0.20.0